Move tree actions logic to the separate component
