+++ src/ggrc-client/js/components/tree/tests/tree-widget-container_spec.js
@@ -4,7 +4,6 @@
  */
 
 import * as TreeViewUtils from '../../../plugins/utils/tree-view-utils';
-import * as SnapshotUtils from '../../../plugins/utils/snapshot-utils';
 import * as CurrentPageUtils from '../../../plugins/utils/current-page-utils';
 import * as AdvancedSearch from '../../../plugins/utils/advanced-search-utils';
 import * as ObjectVersions from '../../../plugins/utils/object-versions-utils';
@@ -104,97 +103,6 @@ describe('GGRC.Components.treeWidgetContainer', function () {
       vm.attr('loaded', dfd);
       result = display();
       expect(result).toBe(dfd);
-    });
-  });
-
-  describe('addItem get() method', function () {
-    describe('if there is options.objectVersion', function () {
-      beforeEach(function () {
-        vm.attr('options.objectVersion', {data: 1});
-      });
-
-      it('returns false', function () {
-        expect(vm.attr('addItem')).toBe(false);
-      });
-    });
-
-    describe('if there is no options.objectVersion', function () {
-      beforeEach(function () {
-        vm.attr('options.objectVersion', null);
-      });
-
-      it('returns options.add_item_view if it exists', function () {
-        let expectedData = new can.Map({});
-        vm.attr('options.add_item_view', expectedData);
-        expect(vm.attr('addItem')).toBe(expectedData);
-      });
-
-      it('returns model.tree_view_options.add_item_view by default',
-        function () {
-          let expectedData = new can.Map({});
-          vm.attr('options.add_item_view', null);
-          vm.attr('model', {
-            tree_view_options: {
-              add_item_view: expectedData,
-            },
-          });
-          expect(vm.attr('addItem')).toBe(expectedData);
-        });
-    });
-  });
-
-  describe('isSnapshot get() method', function () {
-    let isSnapshotScope;
-    let isSnapshotModel;
-
-    beforeEach(function () {
-      isSnapshotScope = spyOn(SnapshotUtils, 'isSnapshotScope');
-      isSnapshotModel = spyOn(SnapshotUtils, 'isSnapshotModel');
-    });
-
-    describe('if parent_instance is a snapshot scope and ' +
-    'model.model_singular is a snapshot model', function () {
-      beforeEach(function () {
-        vm.attr('parent_instance', {data: 'Data'});
-        vm.attr('model', {model_singular: 'modelSingular'});
-
-        isSnapshotScope.and.returnValue({data: '1'});
-        isSnapshotModel.and.returnValue({data: '2'});
-      });
-
-      it('returns true value', function () {
-        expect(vm.attr('isSnapshots')).toBeTruthy();
-        expect(isSnapshotScope).toHaveBeenCalledWith(
-          vm.attr('parent_instance')
-        );
-        expect(isSnapshotModel).toHaveBeenCalledWith(
-          vm.attr('model.model_singular')
-        );
-      });
-    });
-
-    it('returns options.objectVersion by default', function () {
-      vm.attr('options.objectVersion', {data: 'Data'});
-      expect(vm.attr('isSnapshots')).toBeTruthy();
-    });
-
-    describe('if parent_instance is not a snapshot scope or ' +
-    'model.model_singular is not a snapshot model', function () {
-      beforeEach(function () {
-        isSnapshotScope.and.returnValue(null);
-        isSnapshotModel.and.returnValue(null);
-      });
-
-      it('returns true value if there is options.objectVersion', function () {
-        vm.attr('options.objectVersion', {data: 'Data'});
-        expect(vm.attr('isSnapshots')).toBeTruthy();
-      });
-
-      it('returns false value if there is no options.objectVersion',
-        function () {
-          vm.attr('options.objectVersion', null);
-          expect(vm.attr('isSnapshots')).toBeFalsy();
-        });
     });
   });
